Move functionality to a new file
